requirements.txt
Django==2.1.2
djangorestframework==3.9.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.5
six==1.11.0
